requirements.txt
fastapi>=0.100.0
uvicorn==0.27.1
python-multipart==0.0.9
PyMuPDF==1.23.26
sentence-transformers==2.5.1
scikit-learn==1.6.1
python-dotenv==1.0.1
pydantic>=2.0.0
numpy==1.26.4
torch==2.2.1
transformers==4.38.2
google-genai>=0.1.0
sqlalchemy==2.0.27
aiosqlite==0.20.0